+++ src/main/java/Task2/BedSheets.java
@@ -1,0 +1,7 @@
+package Task2;
+
+public class BedSheets extends HomeTextiles{
+    public BedSheets(){
+
+    }
+}
