+++ src/main/java/Task2/Curtains.java
@@ -1,0 +1,7 @@
+package Task2;
+
+public class Curtains extends HomeTextiles{
+    public Curtains(){
+
+    }
+}
