+++ src/main/java/Task2/Mattresses.java
@@ -1,0 +1,10 @@
+package Task2;
+
+public class Mattresses extends HomeTextiles{
+
+    public Mattresses(){
+
+    }
+
+
+}
