+++ src/main/java/Task2/Pillows.java
@@ -1,0 +1,8 @@
+package Task2;
+
+public class Pillows extends HomeTextiles{
+    public Pillows(){
+
+    }
+
+}
