--- task2/BedSheets.java
@@ -1,9 +1,0 @@
-package Task2;
-
-public class BedSheets extends HomeTextiles{
-    public BedSheets(){
-        height = 200;
-        width = 160;
-        type = "Bed Sheets";
-    }
-}
