--- task2/Curtains.java
@@ -1,9 +1,0 @@
-package Task2;
-
-public class Curtains extends HomeTextiles{
-    public Curtains(){
-        super.height = 90;
-        super.width = 150;
-        super.type = "Curtains";
-    }
-}
