--- task2/Mattresses.java
@@ -1,12 +1,0 @@
-package Task2;
-
-public class Mattresses extends HomeTextiles{
-
-    public Mattresses(){
-        super.height = 200;
-        super.width = 140;
-        super.type = "Mattress";
-    }
-
-
-}
