--- task2/Pillows.java
@@ -1,10 +1,0 @@
-package Task2;
-
-public class Pillows extends HomeTextiles{
-    public Pillows(){
-        height = 40;
-        width = 70;
-        type = "Pillow";
-    }
-
-}
